create pet finished

+++ src/useCases/petCases/deletePet.ts
@@ -3,14 +3,22 @@ import { SOMETHING_IS_WRONG, SUCCESS_RESPONSE } from '../../constants/constants'
 import { prisma } from '../../database/prisma'
 
 export const deletePet = async (req: Request, res: Response) => {
-  const { petId, userRole } = req?.query as any
   try {
-    if (userRole === 'VOLUNTEER') {
-      await prisma.petsCaredByVolunteer.deleteMany({
-        where: { petId },
+    const { petId } = req?.params as any
+
+    const pet = await prisma.pet.findUnique({
+      where: { id: petId as string },
+    })
+
+    await prisma.petVaccine.deleteMany({
+      where: { petId: petId as string },
+    })
+
+    if (pet?.createdBy !== (req.user as any)?.id)
+      return res.status(401).json({
+        ok: false,
+        message: 'You can not delete this pet',
       })
-    }
-    console.log('petId', petId)
 
     if (petId)
       await prisma.pet.delete({
